updated to use the new implementation. Add formatDiffs
