+++ src/createSatisfier.ts
@@ -1,147 +1,164 @@
-// import { Diff, Satisfier } from './interfaces'
-// import { ArrayEntryExpectation } from './ArrayEntryExpectation'
-// import { or } from './logical';
+import { anything } from './anything';
+import { Diff, Predicate, Satisfier } from './interfaces';
 
-// /**
-//  * creates a satisfier
-//  * @param expected All properties can be a value which will be compared to the same property in `actual`, RegExp, or a predicate function that will be used to check against the property.
-//  */
-// export function createSatisfier<T = any>(expected: any): Satisfier<T> {
-//   function test(actual: T) {
-//     return exec(actual) === undefined
-//   }
-//   /**
-//    * Check if `actual` satisfies the expected criteria.
-//    */
-//   function exec(actual: T) {
-//     if (Array.isArray(actual)) {
-//       const diff: Diff[] = []
-//       if (Array.isArray(expected)) {
-//         const arrayEntryExps: ArrayEntryExpectation[] = []
-//         const exp = expected.map(e => {
-//           if (arrayEntryExps.length >= 1) {
-//             return or(...arrayEntryExps)(e)
-//           }
+// Workaround for https://github.com/Microsoft/TypeScript/issues/6230
+export interface ExpectationObject { [key: string]: Expectation }
 
-//           if (e instanceof ArrayEntryExpectation) {
-//             arrayEntryExps.push(e)
-//           }
-//           return e
-//         })
-//         let a = 0
-//         exp.forEach((e: any) => {
-//           if (e === undefined) {
-//             a = a + 1
-//             return
-//           }
-//           diff.push(...detectDiff(actual[a], e, [`[${a}]`], a))
-//           a = a + 1
-//         })
-//         if (actual.length > exp.length) {
-//           for (let i = exp.length; i < actual.length; i++) {
-//             diff.push({ path: [`[${i}]`], expected: undefined, actual: actual[i] })
-//           }
-//         }
-//       }
-//       else if (typeof expected === 'function') {
-//         diff.push(...detectDiff(actual, expected))
-//       }
-//       else if (actual.length === 0) {
-//         diff.push({ path: [], expected: expected, actual })
-//       }
-//       else {
-//         actual.forEach((a, i) => {
-//           diff.push(...detectDiff(a, expected, [`[${i}]`], i))
-//         })
-//       }
-//       return diff.length === 0 ? undefined : diff
-//     }
-//     const diff = detectDiff(actual, expected)
-//     return diff.length === 0 ? undefined : diff
-//   }
-//   return {
-//     expected,
-//     test,
-//     exec
-//   }
-// }
+export interface ExpectionArray extends Array<Expectation> { }
 
-// function detectDiff(actual: any, expected: any, path: string[] = [], index?: number) {
-//   const diff: Diff[] = []
-//   const expectedType = typeof expected
-//   if (expectedType === 'function') {
-//     if (!(expected as Function)(actual, index)) {
-//       diff.push({
-//         path,
-//         expected,
-//         actual
-//       })
-//     }
-//   }
-//   else if (expected === undefined) {
-//     return diff
-//   }
-//   else if (expected === null) {
-//     if (expected !== actual)
-//       diff.push({
-//         path,
-//         expected,
-//         actual
-//       })
-//   }
-//   else if (expectedType === 'number' && typeof actual === 'number') {
-//     if (isNaN(expected) && isNaN(actual)) return diff
-//     if (expected !== actual)
-//       diff.push({ path, expected, actual })
-//   }
-//   else if (expectedType === 'boolean' || expectedType === 'number' || expectedType === 'string' || actual === undefined) {
-//     if (expected !== actual)
-//       diff.push({ path, expected, actual })
-//   }
-//   else if (expected instanceof ArrayEntryExpectation) {
-//     const d = expected.exec(actual, path)
-//     if (d) diff.push(...d)
-//   }
-//   else if (expected instanceof RegExp) {
-//     if (!expected.test(actual)) {
-//       diff.push({
-//         path,
-//         expected,
-//         actual
-//       })
-//     }
-//   }
-//   else if (Array.isArray(expected)) {
-//     if (!Array.isArray(actual)) {
-//       diff.push({
-//         path,
-//         expected,
-//         actual
-//       })
-//     }
-//     else {
-//       expected.forEach((e: any, i) => {
-//         if (e === undefined)
-//           return
-//         const actualValue = actual[i]
-//         diff.push(...detectDiff(actualValue, e, path.concat([`[${i}]`]), i))
-//       })
-//     }
-//   }
-//   else {
-//     // expected is object. If actual is not, then it is diff.
-//     const actualType = typeof actual
-//     if (actualType === 'boolean' || actualType === 'string' || actualType === 'number' || actual === undefined || actual === null)
-//       diff.push({
-//         path,
-//         expected,
-//         actual
-//       })
-//     else {
-//       Object.keys(expected).forEach(k => {
-//         diff.push(...detectDiff(actual[k], expected[k], path.concat([k])))
-//       })
-//     }
-//   }
-//   return diff
-// }
+export type Expectation = symbol | undefined | null | boolean | number | bigint |
+  string | RegExp | ExpectionArray | ExpectationObject | Predicate
+
+export function createSatisfier<T = any>(expected: Expectation): Satisfier<T> {
+  return {
+    expected,
+    test(actual: T) {
+      return this.exec(actual) === undefined
+    },
+    exec(actual: T) {
+      const d = diff(expected, actual)
+      return d.length ? d : undefined
+    }
+  }
+}
+
+const nodiff: Diff[] = []
+function diff(expected: any, actual: any, path: Diff['path'] = [], index?: number): Diff[] {
+  if (expected === anything) {
+    return nodiff
+  }
+
+  if (expected === undefined || expected === null) {
+    return actual === expected ? nodiff : [{
+      path,
+      expected,
+      actual
+    }]
+  }
+
+  // tslint:disable-next-line: strict-type-predicates valid-typeof
+  if (typeof expected === 'bigint') {
+    return actual === expected ? nodiff : [{
+      path,
+      expected,
+      actual
+    }]
+  }
+
+  const expectedType = typeof expected
+
+  if (expectedType === 'number') {
+    if (isNaN(expected)) {
+      return typeof actual === 'number' && isNaN(actual) ? nodiff : [{
+        path,
+        expected,
+        actual
+      }]
+    }
+    else {
+      return actual === expected ? nodiff : [{
+        path,
+        expected,
+        actual
+      }]
+    }
+  }
+
+  if (expectedType === 'boolean' || expectedType === 'string' || expectedType === 'symbol') {
+    return actual === expected ? nodiff : [{
+      path,
+      expected,
+      actual
+    }]
+  }
+
+  if (expected instanceof RegExp) {
+    if (actual instanceof RegExp) {
+      return expected.test(String(actual)) ? nodiff : [{
+        path,
+        expected,
+        actual
+      }]
+    }
+    else {
+      return (typeof actual === 'string') && expected.test(actual) ?
+        nodiff :
+        [{
+          path,
+          expected,
+          actual
+        }]
+    }
+  }
+
+  if (Array.isArray(expected)) {
+    if (!Array.isArray(actual)) {
+      return [{
+        path,
+        expected,
+        actual
+      }]
+    }
+    else {
+      const max = Math.max(expected.length, actual.length)
+      const diffs: Diff[] = []
+      for (let i = 0; i < max; i++) {
+        const e = expected[i]
+        const a = actual[i]
+        diffs.push(...diff(e, a, path.concat([i])))
+      }
+
+      return diffs
+    }
+  }
+
+  if (expectedType === 'function') {
+    const r = (expected as Predicate)(actual, path)
+    if (r === true) return nodiff
+    return r ? r : [{
+      path,
+      expected,
+      actual
+    }]
+  }
+
+  // expected is an object
+  if (actual === undefined || actual === null) {
+    return [{
+      path,
+      expected,
+      actual
+    }]
+  }
+  const actualType = typeof actual
+  if (actualType === 'boolean' ||
+    actualType === 'string' ||
+    actualType === 'number' ||
+    actualType === 'bigint' ||
+    actualType === 'symbol' ||
+    actualType === 'function' ||
+    actual instanceof RegExp) {
+    return [{
+      path,
+      expected,
+      actual
+    }]
+  }
+  else if (Array.isArray(actual)) {
+    return [{
+      path,
+      expected,
+      actual
+    }]
+    // return actual.reduce((p, v, i) => {
+    //   p.push(...diff(expected, v, path.concat([i])))
+    //   return p
+    // }, [] as Diff[])
+  }
+  else {
+    return Object.keys(expected).reduce((p, key: string) => {
+      p.push(...diff(expected[key], actual[key], path.concat([key])))
+      return p
+    }, [] as Diff[])
+  }
+}
