+++ src/every.ts
@@ -5,7 +5,7 @@ import {
   Tersible
 } from 'tersify'
 
-import { createSatisfier } from './createSatisfier2'
+import { createSatisfier } from './createSatisfier'
 
 /**
  * Check if every entry in the array satisfies the expectation.
@@ -13,5 +13,12 @@ import { createSatisfier } from './createSatisfier2'
  */
 export function every(expectation: any) {
   const s = createSatisfier(expectation)
-  return tersible((e: any) => e && Array.isArray(e) && e.every(v => s.test(v)), () => `every(${tersify(expectation)})`)
+  return tersible((e: any) => e && Array.isArray(e) && e.reduce((p, v, i) => {
+    const d = s.exec(v)
+    if (d) p.push(...d.map(d => {
+      d.path.unshift(i)
+      return d
+    }))
+    return p
+  }, []), () => `every(${tersify(expectation)})`)
 }
