+++ src/has.ts
@@ -5,7 +5,7 @@ import {
   Tersify
 } from 'tersify'
 
-import { createSatisfier } from './createSatisfier2'
+import { createSatisfier } from './createSatisfier'
 
 /**
  * Check if an array has entries satisfy the expectations in order.
@@ -14,7 +14,7 @@ export function has(...expectations: any[]) {
   return tersible((arr: any) => {
     if (!Array.isArray(arr)) return false
     let index = 0
-    return expectations.some(e => {
+    return expectations.every(e => {
       const s = createSatisfier(e)
       let actual = arr[index]
       while (index < arr.length && !s.test(actual)) {
