+++ src/index.ts
@@ -1,13 +1,15 @@
 export * from './anything';
-export * from './ArrayEntryExpectation';
-export * from './createSatisfier2';
+// export * from './ArrayEntryExpectation';
+export * from './createSatisfier';
 export * from './every';
+// export * from './and'
+// export * from './or';
+export * from './formatDiffs';
 export * from './has';
 export * from './interfaces';
 export * from './isInInterval';
 export * from './isInRange';
 export * from './isTypeOf';
 export * from './none';
-// export * from './or';
 export * from './some';
-// export * from './and'
+export * from './startsWith';
