+++ src/some.ts
@@ -5,7 +5,7 @@ import {
   Tersible
 } from 'tersify'
 
-import { createSatisfier } from './createSatisfier2'
+import { createSatisfier } from './createSatisfier'
 
 /**
  * Check if an array have at least one entry satisfying the expectation.
